+++ src/build_number.h
@@ -1,1 +1,1 @@
-#define BUILD 525
+#define BUILD 526