+++ src/services/scriptingservice.h
@@ -65,6 +65,9 @@ public:
     Q_INVOKABLE QString noteTextEditSelectedText();
     Q_INVOKABLE void noteTextEditSelectAll();
     Q_INVOKABLE void noteTextEditSelectCurrentLine();
+    Q_INVOKABLE void noteTextEditSetSelection(int start, int end);
+    Q_INVOKABLE int noteTextEditSelectionStart();
+    Q_INVOKABLE int noteTextEditSelectionEnd();
     Q_INVOKABLE QString noteTextEditCurrentWord(
             bool withPreviousCharacters = false);
     Q_INVOKABLE void encryptionDisablePassword();